requirements.txt
aiohappyeyeballs==2.4.0
aiohttp>=3.8.4
aiohttp-retry>=2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs>=19.3,<22.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.2
fastapi==0.100.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==5.2.0
pydantic==1.10.9
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
twilio==8.2.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
websockets>=10.4
yarl==1.8.2
sanic==21.12.0
PyYAML==5.4.1
rasa==3.6.2
